클라이언트 수정 중

+++ src/main/java/service/UserService.java
@@ -32,7 +32,7 @@ public class UserService {
         this.detailsDAO = detailsDAO;
     }
 
-    public Long insertUser(UserDTO user) {
+    public int insertUser(UserDTO user) {
         return userDAO.insertUser(user);
     }
